HW15: Rename generated package name by sqlc
