+++ hw15_go_sql/internal/db/db.go
@@ -2,7 +2,7 @@
 // versions:
 //   sqlc v1.28.0
 
-package repository
+package db
 
 import (
 	"context"
