+++ hw15_go_sql/internal/db/querier.go
@@ -2,7 +2,7 @@
 // versions:
 //   sqlc v1.28.0
 
-package repository
+package db
 
 import (
 	"context"
