+++ hw15_go_sql/internal/db/query.sql.go
@@ -3,7 +3,7 @@
 //   sqlc v1.28.0
 // source: query.sql
 
-package repository
+package db
 
 import (
 	"context"
