+++ hw15_go_sql/sqlc.yaml
@@ -5,8 +5,8 @@ sql:
     engine: "postgresql"
     gen:
       go:
-        package: "repository"
-        out: "internal/repository"
+        package: "db"
+        out: "internal/db"
         sql_package: "pgx/v5"
         emit_interface: true
         emit_db_tags: true
